Renames build2 project to linc
Previous name (line-collision-detector) was also the name of
the git repo, and it was too long to both type and read.
Particularly since the build2 canonical project structure
dictates that the source directory and the binary are both
named like the build2 project itself.
That gave us 4 things in total that were all called
line-collision-detector.

Now we get one thing called line-collision-detector,
and three things called linc.